allow member search to display more than 100 results
Previously only the first 100 results would show on member searches.
This wasn't very useful, and paging results didn't seem helpful either
as it makes it hard to do other things (like grep for example) to the
results.

So now we just run the member_search command multiple times and collect
results to show at once.

Also, found that table_print is terrible for bigger tables, so for this
output I do the table output by hand.